Move text settings into a separate folder in prep for refactoring
